Trekk ut businesslogikk i egen hook
