+++ src/frontend/assets/lang/nb.json
@@ -22,6 +22,10 @@
   "dokumentasjon.etter-bilde.punktliste": "<punktliste>\n<punkt> Dokumentet har riktig retning.</punkt>\n<punkt> Teksten til dokumentet er godt leselig.</punkt>\n<punkt> Bildet er godt opplyst, uten skygger.</punkt>\n</punktliste>",
   "dokumentasjon.har-sendt-inn.spm": "Jeg har sendt inn denne dokumentasjonen til NAV tidligere",
   "dokumentasjon.info": "Du må laste opp dokumentasjon som bekrefter opplysningene du har gitt i søknaden. Du får raskere svar på søknaden hvis vi har all dokumetasjon når vi starter behandlingen. Du kan bruke filformatene PDF, jpg og png. {linjeskift} {linjeskift} Slik bruker du bilder som vedlegg til søknaden:{linjeskift} {linjeskift} 1. Ta bilde av dokumentet med smarttelefon eller nettbrett {linjeskift} 2. Sjekk at dokumentet er lett å lese {linjeskift} 3. Last opp bildene her {linjeskift} {linjeskift} Mangler du noe av dokumentasjonen? Da kan du sende inn det du har og ettersende resten. {linjeskift} {linjeskift} Har du sendt inn denne dokumentasjonen til NAV tidligere? Da behøver du ikke å sende den på nytt.",
+  "dokumentasjon.last-opp-dokumentasjon.aktivknapp": "Slipp filen her...",
+  "dokumentasjon.last-opp-dokumentasjon.feilmeldinggenerisk": "Noe gikk feil da vi prøvde å laste opp filen din. Prøv igjen! Fil: {filnavn}",
+  "dokumentasjon.last-opp-dokumentasjon.feilmeldingstor": "Filen du prøvde å laste opp er for stor. Maksimal størrelse er {maks}. Fil: {filnavn}",
+  "dokumentasjon.last-opp-dokumentasjon.feilmeldingtype": "Filen du prøvde å laste opp har feil filtype. Godkjente filtyper er png, pdf, jpg og jpeg. Fil: {filnavn}",
   "dokumentasjon.last-opp-dokumentasjon.knapp": "Last opp dokumentasjon",
   "dokumentasjon.mer-hjelp.lenke": "https://www.nav.no/no/nav-og-samfunn/kontakt-nav/relatert-informasjon/ta-bilde-av-vedleggene-med-mobilen",
   "dokumentasjon.mer-hjelp.lenketekst": "Mer hjelp til å laste opp vedlegg (åpnes i nytt vindu).",
@@ -44,8 +48,8 @@
   "felles.fødselsdato.label": "Fødselsdato (format dd.mm.åååå)",
   "felles.mangler-svar.feilmelding": "Du må svare på dette spørsmålet for å gå videre",
   "felles.modal.avbrytogslett.brødtekst": "Det du har fylt ut i søknaden blir slettet, og du kommer tilbake til velkomstsiden.{linjeskift]{linjeskift]Ønsker du å slette søknaden?",
-  "felles.modal.avbrytogslett.ja": "JA, AVBRYT SØKNAD",
-  "felles.modal.avbrytogslett.tittel": "Avbryt og slett søknad",
+  "felles.modal.avbrytogslett.ja": "JA, AVBRYT SØKNADEN",
+  "felles.modal.avbrytogslett.tittel": "Avbryt og slett søknaden",
   "felles.navigasjon.avsluttogfortsett": "Avslutt og fortsett senere",
   "felles.navigasjon.avsluttogslett": "Avbryt og slett søknad",
   "felles.navigasjon.gå-videre": "GÅ VIDERE",
@@ -153,8 +157,8 @@
   "ombarnet.andre-forelder.navn-ukjent.spm": "Jeg kan ikke gi opplysninger om den andre forelderen",
   "ombarnet.andre-forelder.navn.feilmelding": "Du må opplyse om den andre forelderens navn for å gå videre",
   "ombarnet.andre-forelder.navn.spm": "Navn ",
-  "ombarnet.andre-forelder.utenlandspensjon.spm": "Får den andre forelderen til {navn} pensjon fra utlandet?",
   "ombarnet.andre-forelder.utenlandspensjon.land.spm": "Hvilket land får den andre forelderen pensjon fra?",
+  "ombarnet.andre-forelder.utenlandspensjon.spm": "Får den andre forelderen til {navn} pensjon fra utlandet?",
   "ombarnet.barnetrygd-eøs": "Du har opplyst at du får eller har søkt om barnetrygd\nfor {navn} fra et annet EU/EØS land.",
   "ombarnet.barnetrygd-eøs.land.spm": "Hvilket land får eller har du søkt om barnetrygd fra?",
   "ombarnet.bor-fast.spm": "Bor {navn} fast sammen med deg?",
@@ -202,7 +206,7 @@
   "omdeg.borpådenneadressen.ikke-endre-adresse": "Skal du ikke endre adresse i Folkeregisteret?",
   "omdeg.borpådenneadressen.kontakt-folkeregister.alert": "Du må kontakte Folkeregisteret og endre til riktig adresse.",
   "omdeg.borpådenneadressen.spm": "Bor du på denne adressen?",
-  "omdeg.borpådenneadressen.spm.tilleggsinfo": "Adressen vises ikke her hvis du har strengt fortrolig adresse.",
+  "omdeg.borpådenneadressen.spm.tilleggsinfo": "Adressen vises ikke her hvis du har fortrolig eller strengt fortrolig adresse.",
   "omdeg.du-kan-ikke-søke.feilmelding": "Du må bruke PDF-skjema for å søke dersom du ikke skal endre adresse i Folkeregisteret",
   "omdeg.endre-opplysninger.lenke": "https://www.skatteetaten.no/person/folkeregister/endre/",
   "omdeg.endre-opplysninger.lenketekst": "Endre opplysninger i Folkeregisteret (åpnes i nytt vindu)",
@@ -226,6 +230,7 @@
   "omdeg.telefon.spm": "Telefonnummer du kan kontaktes på",
   "omdeg.utenlandspensjon.land.spm": "Hvilket land får du pensjon fra?",
   "omdeg.utenlandspensjon.spm": "Får du pensjon fra utlandet?",
+  "oppsummering.deltittel.ombarnet": "{nummer}. Om {navn}",
   "oppsummering.endresvar.lenketekst": "Endre svarene dine",
   "oppsummering.info": "Les oppsummeringen nøye før du sender inn søknaden. {linjeskift} Hvis du vil gjøre endringer, kan du gå tilbake.",
   "oppsummering.sidetittel": "Oppsummering"
